+++ configs/lisa_pico.py
@@ -37,7 +37,10 @@ class ModulationSection(nallely.Module):
 
 
 class WavetableSection(nallely.Module):
-    stream_input = nallely.ModulePitchwheel(stream=True)
+    stream_table1 = nallely.ModulePitchwheel(stream=True, channel=0)
+    stream_table2 = nallely.ModulePitchwheel(stream=True, channel=1)
+    stream_table3 = nallely.ModulePitchwheel(stream=True, channel=2)
+    stream_table4 = nallely.ModulePitchwheel(stream=True, channel=3)
     capture_mode = nallely.ModuleParameter(122, description='Selects the type of capture mode', accepted_values=['forward', 'backward', 'rolling', 'buf-forward-rl', 'buf-backward-lr', 'rolling-backward'])
     phase_reset = nallely.ModuleParameter(126, description='Reset the phase')
     phase_offset = nallely.ModuleParameter(125, description='Add an offset to the phase')